Make update user generic function in helper file and move all logic into it.
