+++ routes/user.js
@@ -6,8 +6,6 @@ const router = express.Router();
 
 router.get("/user", user.getUser);
 router.put("/user/:id", user.updateUser);
-router.put("/user/update-third-party-user/:id", user.updateThirdPartyUser);
-router.put("/user/update-password/:id", user.updateUserPassword);
 router.delete("/user/:id", user.deleteUser);
 
 export default router;
